chore(pacquet-cli): split list tests into separate test file
